Split dumbbell series to multiple files
