+++ ts/Series/Dumbbell/DumbbellPointOptions.d.ts
@@ -1,0 +1,38 @@
+/* *
+ *
+ *  (c) 2010-2020 Torstein Honsi
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type AreaRangePointOptions from '../AreaRange/AreaRangePointOptions';
+import type ColorType from '../../Core/Color/ColorType';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+interface DumbbellPointOptions extends AreaRangePointOptions {
+    connectorColor?: ColorType;
+    connectorWidth?: number;
+    dashStyle?: string;
+    lowColor?: ColorType;
+}
+
+/* *
+ *
+ *  Default export
+ *
+ * */
+
+export default DumbbellPointOptions;
